refactor: manage.js를 manager.js로 변경
